% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chembl_moa_list}
\alias{chembl_moa_list}
\title{ChEMBL Mechanism-of-Action Annotations}
\format{
A data frame containing drug identifiers and their mechanism or target annotations.
}
\description{
Drug and mechanism-of-action annotations derived from ChEMBL.
}
