% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrustRank.R
\name{createGraphNetwork}
\alias{createGraphNetwork}
\title{Create Complete Graph Network}
\usage{
createGraphNetwork(
  drug_node_edge,
  gene_node_edge,
  drug_node,
  gene_node,
  file_name
)
}
\arguments{
\item{drug_node_edge}{Drug node-to-edge mappings.}

\item{gene_node_edge}{Gene node-to-edge mappings.}

\item{drug_node}{Drug node data.}

\item{gene_node}{Gene node data.}

\item{file_name}{Output file name for the graph network.}
}
\description{
Integrates PPI and drug-target graphs into a complete network.
}
