% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_seeds}
\alias{disease_seeds}
\title{Disease Seed Genes}
\format{
A data frame with one column:
\describe{
  \item{gene}{Ensembl gene identifier of a SARS-CoV-2 host protein target.}
}
}
\source{
Gysi DM, do Valle I, Zitnik M, et al. (2021).
Network medicine framework for identifying drug-repurposing opportunities
for COVID-19.
\emph{Proceedings of the National Academy of Sciences}, 118(19),
e2025581118.
\doi{10.1073/pnas.2025581118}
}
\usage{
disease_seeds
}
\description{
An example set of disease-associated seed genes for demonstrating
network-based drug repurposing with DrugSigNet.
}
\details{
The seed genes are derived from the SARS-CoV-2 host protein targets used by
Gysi et al. (2021) to define the COVID-19 disease module for network-based
drug repurposing. The original set comprised 332 experimentally identified
human proteins targeted by SARS-CoV-2 proteins. Gene symbols were mapped to
Ensembl gene identifiers for use with the DrugSigNet network-based workflow.
}
\examples{
data(disease_seeds)
head(disease_seeds)

}
\keyword{datasets}
