% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{disease_signature}
\alias{disease_signature}
\title{Disease Gene-Expression Signature}
\format{
A data frame with two columns:
\describe{
  \item{Entrez}{NCBI Entrez Gene identifier.}
  \item{FC}{Log2 fold-change for COVID-19 convalescent donors relative to
  healthy donors. Positive and negative values indicate higher and lower
  expression in convalescent donors, respectively.}
}
}
\source{
Gedda MR, Danaher P, Shao L, et al. (2022).
Longitudinal transcriptional analysis of peripheral blood leukocytes in
COVID-19 convalescent donors.
\emph{Journal of Translational Medicine}, 20, 587.
\doi{10.1186/s12967-022-03751-7}
}
\usage{
disease_signature
}
\description{
An example disease gene-expression signature for demonstrating
signature-based drug repurposing with DrugSigNet.
}
\details{
The signature is derived from peripheral blood leukocyte transcriptomic
profiles of COVID-19 convalescent donors (CCD) compared with healthy donors
(HD), as reported by Gedda et al. (2022). Genes from the 120--149 day
post-symptom-onset comparison were retained when the false discovery rate
(FDR) was less than or equal to 0.05 and a log2 fold-change value was
available. Gene identifiers were mapped to NCBI Entrez Gene identifiers.
}
\examples{
data(disease_signature)
head(disease_signature)

}
\keyword{datasets}
