% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRA.R
\name{.transform_to_ranks}
\alias{.transform_to_ranks}
\title{Helper Function to Compute Ranks for a Vector}
\usage{
.transform_to_ranks(x, ties_method, reverse)
}
\arguments{
\item{x}{A numeric vector to rank.}

\item{ties_method}{A character string specifying the method to handle ties ("average", "first", "last", "random", "dense").}

\item{reverse}{A logical indicating if rankings should be reversed.}
}
\value{
A numeric vector of ranks with ties handled and optional reversal applied.
}
\description{
This function computes ranks for a numeric vector, with optional tie-breaking methods and reverse ordering.
}
\keyword{internal}
