% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dowdall.R
\name{Dowdall,RankAggregation-method}
\alias{Dowdall,RankAggregation-method}
\title{Dowdall for a RankAggregation Object}
\usage{
\S4method{Dowdall}{RankAggregation}(object)
}
\arguments{
\item{object}{A \code{RankAggregation} object containing the input rankings and Dowdall parameters.}
}
\value{
The supplied \code{RankAggregation} object with aggregation results stored in its \code{result} slot.
}
\description{
Dowdall for a RankAggregation Object
}
