% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DrugSearchingPipeline-class}
\alias{DrugSearchingPipeline-class}
\title{DrugSearchingPipeline Class Definition}
\description{
A class to manage drug signature searching and processing.
}
