% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{DrugSearchingPipelineVisualized-class}
\alias{DrugSearchingPipelineVisualized-class}
\title{Visualized Drug-Searching Pipeline Class}
\description{Stores raw and processed searches, rank aggregations, and visualizations.}
\section{Slots}{\describe{
\item{\code{DrugSearching}}{A list containing \code{Raw} and \code{Processed} search results.}
\item{\code{RankAggregation}}{A list of rank-aggregation results.}
\item{\code{Visualization}}{A list of visualization results.}
\item{\code{type}}{A character string identifying the pipeline type.}
}}
