% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_utils.R
\name{format_pipeline_task_label}
\alias{format_pipeline_task_label}
\title{Format a pipeline task label}
\usage{
format_pipeline_task_label(task, idx, task_label = NULL)
}
\description{
Format a pipeline task label
}
\keyword{internal}
