% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrustRank.R
\name{generateGraphFiles}
\alias{generateGraphFiles}
\title{Generate Graph Files}
\usage{
generateGraphFiles(temp_dir, ppi_network, drug_target_network, disease_genes)
}
\arguments{
\item{temp_dir}{Temporary directory for storing intermediate files.}

\item{ppi_network}{Protein-protein interaction network. Expected to contain
columns `gene1` and `gene2`. If `NULL`, the default `"gene_gene"` network
is loaded.}

\item{drug_target_network}{Drug-target interaction network. Expected to
contain columns `ID`, `Drug`, `Target`, and `Group`. If `NULL`, the default
`"drug_target"` network is loaded.}

\item{disease_genes}{Data frame containing disease seed genes in a column
named `gene`.}
}
\value{
A list of file paths for the graph components.
}
\description{
Generates the necessary files for running the TrustRank algorithm.
}
