% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_utils.R
\name{left_join_annotations_by_drug}
\alias{left_join_annotations_by_drug}
\title{Join annotations to an input drug table}
\usage{
left_join_annotations_by_drug(input_tbl, ref_tbl, empty_cols = character())
}
\arguments{
\item{input_tbl}{Input table with `Drug` column.}

\item{ref_tbl}{Reference table with `Drug` column.}

\item{empty_cols}{Annotation columns to include when `ref_tbl` is empty.}
}
\value{
Joined data frame.
}
\description{
Join annotations to an input drug table
}
\keyword{internal}
