% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lincs_pert_info}
\alias{lincs_pert_info}
\title{LINCS Perturbagen Metadata}
\format{
A data frame with one row per perturbagen and identifier and annotation columns.
}
\description{
Perturbagen annotations for the bundled LINCS reference data.
}
