% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_proximity.R
\name{Network_proximity,NetworkBased-method}
\alias{Network_proximity,NetworkBased-method}
\title{Network Proximity Method for Network-Based Drug Search}
\usage{
\S4method{Network_proximity}{NetworkBased}(object)
}
\arguments{
\item{object}{A \code{NetworkBased} object containing the networks, disease genes, and network-proximity parameters.}
}
\value{
The same \code{NetworkBased} object with the results of the Network Proximity analysis.
}
\description{
Method for the Network Proximity drug search using a network-based proximity algorithm.
The method performs simulations and ranks drugs based on their proximity to disease genes.
}
