% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NetworkBased-class}
\alias{NetworkBased-class}
\title{Create a SignatureBasedDrugSearching Object}
\arguments{
\item{signature_result}{A `data.frame` containing the results of the search.}

\item{query}{A `list` containing the up-regulated (`upset`) and down-regulated (`downset`) gene sets.}

\item{signature_method}{A `character` string representing the method used (e.g., "CMAP").}

\item{refdb}{A `character` string specifying the reference database used (e.g., "cmap").}
}
\value{
A `SignatureBasedDrugSearching` object.
}
\description{
A helper function to create an instance of the `SignatureBasedDrugSearching` class.
}
