% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrustRank.R
\name{validateInputs}
\alias{validateInputs}
\title{Validate Inputs}
\usage{
validateInputs(object)
}
\arguments{
\item{object}{Optional `NetworkBased` object. If `NULL`, a new object is
created from the supplied network inputs and method parameters.}
}
\value{
Throws an error if inputs are invalid, otherwise invisible.
}
\description{
Validates the required inputs for the TrustRank analysis.
}
